* Correction of the repos structure for addonTemplate so that the add-on can be built with the scons command.
