+++ manifest-translated.ini.tpl
@@ -1,0 +1,2 @@
+summary = "{addon_summary}"
+description = """{addon_description}"""
